Revert consolidation of assert_*_regex methods
